refactor the stateless-core
