+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "bin/stateless-validator",
     "crates/stateless-common",
     "crates/stateless-core",
+    "crates/stateless-db",
 ]
 resolver = "2"
