+++ bin/stateless-validator/Cargo.toml
@@ -29,8 +29,10 @@ salt.workspace = true
 # stateless
 stateless-common = { path = "../../crates/stateless-common" }
 stateless-core = { path = "../../crates/stateless-core" }
+stateless-db = { path = "../../crates/stateless-db" }
 
 # misc
+alloy-consensus.workspace = true
 bincode.workspace = true
 clap = { workspace = true, features = ["env"] }
 eyre.workspace = true
@@ -39,6 +41,7 @@ kanal.workspace = true
 metrics.workspace = true
 metrics-exporter-prometheus.workspace = true
 num_cpus.workspace = true
+redb.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tokio.workspace = true
