+++ crates/stateless-db/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "stateless-db"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+[dependencies]
+# alloy
+alloy-genesis.workspace = true
+alloy-primitives.workspace = true
+alloy-rpc-types-eth.workspace = true
+
+# op
+op-alloy-rpc-types.workspace = true
+
+# revm
+revm.workspace = true
+
+# stateless
+stateless-core = { path = "../stateless-core" }
+
+# misc
+bincode.workspace = true
+dashmap.workspace = true
+eyre.workspace = true
+lz4_flex.workspace = true
+redb.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+
+[dev-dependencies]
+tempfile = "3"
